qaseio,qase-robotframework: enable Python 3.7
Allow using packages qaseio and qase-robotframework with Python 3.7.
This will enable our clients, still using it, to
receive the latest package updates for Qase reporters.

Change required downgrading several packages to their latest versions
still supporting Python 3.7:
* setuptools from 69 to 68
* typing-extensions from 4.9.0 to 4.7.1

Same change is already made in the main branch, see
commit 946521e381ea2a59daa9ced6962e3758e1287eaa.
In this branch, it's made not for releasing packages,
but to fix builds on Python 3.7.